refactor: Move Status out to its own file
