+++ src/Status.ts
@@ -1,0 +1,10 @@
+/**
+ * Collection of status types supported by the plugin.
+ * TODO: Make this a class so it can support other types and easier mapping to status character.
+ * @export
+ * @enum {number}
+ */
+export enum Status {
+    TODO = 'Todo',
+    DONE = 'Done',
+}
