+++ tests/Query/Filter/StatusField.test.ts
@@ -2,7 +2,7 @@ import { StatusField } from '../../../src/Query/Filter/StatusField';
 import type { FilterOrErrorMessage } from '../../../src/Query/Filter/Filter';
 import { TaskBuilder } from '../../TestingTools/TaskBuilder';
 import { testFilter } from '../../TestingTools/FilterTestHelpers';
-import { Status } from '../../../src/Task';
+import { Status } from '../../../src/Status';
 import {
     expectTaskComparesAfter,
     expectTaskComparesBefore,
